remove telemetry v1
