+++ haystack/__init__.py
@@ -29,6 +29,6 @@ set_pytorch_secure_model_loading()
 import os
 
 
-from haystack.telemetry_2 import send_event
+from haystack.telemetry import send_event
 
 send_event(event_name="Haystack imported")
