+++ haystack/nodes/retriever/_openai_encoder.py
@@ -12,7 +12,7 @@ from haystack.environment import HAYSTACK_REMOTE_API_TIMEOUT_SEC
 from haystack.nodes.retriever._base_embedding_encoder import _BaseEmbeddingEncoder
 from haystack.schema import Document
 from haystack.utils.openai_utils import USE_TIKTOKEN, count_openai_tokens, load_openai_tokenizer, openai_request
-from haystack.telemetry_2 import send_event
+from haystack.telemetry import send_event
 
 if TYPE_CHECKING:
     from haystack.nodes.retriever import EmbeddingRetriever
