+++ haystack/utils/import_utils.py
@@ -11,8 +11,6 @@ from pathlib import Path
 
 import requests
 
-from haystack.telemetry import send_tutorial_event
-
 
 logger = logging.getLogger(__name__)
 
@@ -85,9 +83,6 @@ def fetch_archive_from_http(
     path = Path(output_dir)
     if not path.exists():
         path.mkdir(parents=True)
-
-    if "deepset.ai-farm-qa/datasets" in url or "dl.fbaipublicfiles.com" in url or "fandom-qa.s3" in url:
-        send_tutorial_event(url=url)
 
     is_not_empty = len(list(Path(path).rglob("*"))) > 0
     if is_not_empty:
